refactor: use __path helper for safty check/create  paths
