+++ src/fs/copy.js
@@ -1,10 +1,9 @@
-import { dirname } from 'path';
 import { cp } from 'node:fs/promises';
-import isPathExists from './helpers/isPathExists.js';
+import { extendedDirname, isPathExists } from '../helpers/__path.js';
 
 const copy = async () => {
-  const dirSource = `${dirname(import.meta.filename)}/files`;
-  const dirDestination = `${dirname(import.meta.filename)}/files_copy`;
+  const dirSource = extendedDirname(import.meta.url, 'files');
+  const dirDestination = extendedDirname(import.meta.url, 'files_copy');
 
   if ((await isPathExists(dirDestination)) || !(await isPathExists(dirSource))) {
     throw new Error('FS operation failed');
