+++ src/fs/create.js
@@ -1,9 +1,8 @@
 import { writeFile } from 'node:fs/promises';
-import { dirname } from 'path';
-import isPathExists from './helpers/isPathExists.js';
+import { extendedDirname, isPathExists } from '../helpers/__path.js';
 
 const create = async () => {
-  const filename = `${dirname(import.meta.filename)}/files/freshFile.txt`;
+  const filename = extendedDirname(import.meta.url, 'files', 'freshFile.txt');
 
   if (await isPathExists(filename)) {
     throw new Error('FS operation failed');
