--- src/fs/helpers/isPathExists.js
@@ -1,10 +1,0 @@
-import { access, constants } from 'node:fs/promises';
-
-export default async (filename) => {
-  try {
-    await access(filename, constants.F_OK);
-    return true;
-  } catch {
-    return false;
-  }
-}
