+++ src/fs/list.js
@@ -1,9 +1,9 @@
 import { readdir } from 'node:fs/promises';
-import { dirname } from 'path';
+import { extendedDirname } from '../helpers/__path.js';
 
 const list = async () => {
 	try {
-		const files = await readdir(`${dirname(import.meta.filename)}/files`);
+		const files = await readdir(extendedDirname(import.meta.url, 'files'));
 
 		for (const file of files)
 			console.log(file);
