+++ src/modules/esm.mjs
@@ -2,7 +2,9 @@ import { createRequire } from "module";
 import { dirname, sep } from 'node:path';
 import { release, version } from 'node:os';
 import { createServer as createServerHttp } from 'node:http';
+import { fileURLToPath } from "node:url";
 import './files/c.js';
+import { extendedDirname } from "../helpers/__path.js";
 
 const require = createRequire(import.meta.url);
 const random = Math.random();
@@ -10,17 +12,18 @@ const random = Math.random();
 let unknownObject;
 
 if (random > 0.5) {
-  unknownObject = require(`${dirname(import.meta.filename)}/files/a.json`);
+  // unknownObject = require(`${dirname(import.meta.filename)}/files/a.json`);
+  unknownObject = require(extendedDirname(import.meta.url, 'files', 'a.json'));
 } else {
-  unknownObject = require(`${dirname(import.meta.filename)}/files/b.json`);
+  unknownObject = require(extendedDirname(import.meta.url, 'files', 'b.json'));
 }
 
 console.log(`Release ${release()}`);
 console.log(`Version ${version()}`);
 console.log(`Path segment separator is "${sep}"`);
 
-console.log(`Path to current file is ${import.meta.filename}`);
-console.log(`Path to current directory is ${dirname(import.meta.filename)}`);
+console.log(`Path to current file is ${fileURLToPath(import.meta.url)}`);
+console.log(`Path to current directory is ${extendedDirname(import.meta.url)}`);
 
 const myServer = createServerHttp((_, res) => {
   res.end('Request accepted');
